+++ Kame/src/Kame/Platform/DirectX12/Graphics/TextureDx12.h
@@ -66,7 +66,7 @@ namespace Kame {
 
     virtual void LoadFromFile(const std::wstring& fileName, TextureUsage textureUsage) override;
 
-    virtual Reference<Texture> ToCubeMap(UINT64 width, std::wstring name) override;
+    virtual Reference<Texture> ToCubeMap(UINT width, std::wstring name) override;
 
     TextureUsage GetTextureUsage() const {
       return m_TextureUsage;
